Move handling connection to new service
