+++ src/main/java/com/nokia/uwr/service/ConnectionHandler.java
@@ -1,0 +1,29 @@
+package com.nokia.uwr.service;
+
+import com.nokia.uwr.scenario.bts.BTSDescription;
+
+import java.util.List;
+
+/**
+ * Handler for starting and terminating connection between network-simulator and calls-management-system
+ *
+ * @author Piotr Stoklosa
+ */
+public interface ConnectionHandler {
+
+    /**
+     * Create connection between network-simulator and calls-management-system
+     *
+     * @param descriptionList descriptionList needed to properly initialize board on calls-management-system side.
+     * @author Piotr Stoklosa
+     */
+    void initializeCall(List<BTSDescription> descriptionList);
+
+    /**
+     * Terminate connection between network-simulator and calls-management-system
+     *
+     * @author Piotr Stoklosa
+     */
+    void terminateCall();
+
+}
